requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
Flask-Bootstrap
python-dotenv
email-validator
Flask-Migrate
openai
anthropic 